Fix region_size calculation in bytes
_end and _bottom are HeapWord pointers, while HeapWord is defined as a
HeapWordImpl pointer. Since pointer arithmetic is measured in the size
of the pointed element, '_end - _bottom' is measured in how many
HeapWord elements are needed to fill that gap. sizeof(HeapWord) is 8
bytes in 64-bit architectures, given HeapWord is a pointer itself.

This results in region_size being calculated 8 times lower than its real
value, and the corresponding mapped memory to be much smaller than the
region we are moving, leading to some crashes when allocating objects
through the thread-local allocation buffer (TLAB).