Created a region address map (#8)
* Fixed a bug related to register clobbering and renamed the mapper files (#6)

* Moving a region to a new location (#1)

* experimenting with allocating a new place for a region

* fix

* moving in g1_mark_as_young

* implemented basic mapper

* implemented basic mapper in more eloquent way

* removed comment and breakpoint, this is done

* compressedOops, not applicable

* the current state of my environment

* Protected SanitizeGC code under flag 'SanitizeGC'.

Moved test to a new location.

Reverted innocuous changes in parenthesis wrapping expressions.

* Calling runtime addr mapper from barrier - Dev iteration.

* updating with changes, extending customMapper (#2)

* took the returned mapped value from make_leaf_call and tried to pass that

* Fix region_size calculation in bytes

_end and _bottom are HeapWord pointers, while HeapWord is defined as a
HeapWordImpl pointer. Since pointer arithmetic is measured in the size
of the pointed element, '_end - _bottom' is measured in how many
HeapWord elements are needed to fill that gap. sizeof(HeapWord) is 8
bytes in 64-bit architectures, given HeapWord is a pointer itself.

This results in region_size being calculated 8 times lower than its real
value, and the corresponding mapped memory to be much smaller than the
region we are moving, leading to some crashes when allocating objects
through the thread-local allocation buffer (TLAB).

* Simplify region sizes and limits calculation

region_size is already available as GrainBytes. The mmap() result can be
directly stored as HeapWord* new_bottom. The region size in words is
already available as GrainWords (_end - _bottom pointer difference).
Some auxiliary variables aren't strictly needed.

* Reduce diff against jdk-23-ga

The modified code seems to be there for 'back and forth' reasons, or
because of missing includes in customMapper.hpp.

* Introduce SanitizerGCMapper::remapAddress

This template inline method avoids duplicating the logic to check
whether the SanitizeGC feature is enabled and to remap the address to
its original region.

* Add more SanitizerGCMapper::remapAddress usages

* Add another SanitizerGCMapper::remapAddress usage

This one requires dealing with a const_cast at the same time as the
reinterpret_cast, so I've chosen to use a C-style cast.

* Remove unused beforeEndAddr

* Record mapping as SanitizerGCMapper static fields

Instead of using global variables, introduce SanitizerGCMapper static
fields for SanitizerGCMapper beforeAddr/afterAddr/afterEndAddr. This
makes the SanitizerGCMapper code more organized/scoped.

* Rename static fields

Give more context in terms of the original and moved region boundaries:
 * beforeAddr is the originalRegionStart
 * afterAddr is the movedRegionStart
 * afterEndAddr is the movedRegionEnd

* Compute the moved region offset only once

We don't need to calculate this value on each mapNewAddrToOriginalAddr
call. Instead, we can calculate this in initializeMapping and store its
final value. This works as a minor optimization.

* GC verify debugging (#4)

* update with changes done in the past few weeks, maily remapping from new to original addresses

* fixed a bug related to G1HeapRegion's parsable area

* added the printMemoryRegionMap() function for debugging and other debugging prints

* cleaned up the code, removed some unnecessary assert skipping and added some new mappings

* fixed a bug related to register clobbering, it now pushes and pops registers around the remapping call

* refactoring - renamed customMapper to sanitizeAddressMapper

---------

Co-authored-by: Martin Balao <mbalao@redhat.com>
Co-authored-by: Francisco Ferrari Bihurriet <fferrari@redhat.com>

* removed TODO comment

* refactoring - renamed SanitizerGCMapper class to SanitizeGCMapper

* moving more regions instead of just the first one, created sanitizeHeapRegionList for that

* Started with the implementation of a hash table for heap regions

* creating a map structure for saving and getting moved heap region's addresses

---------

Co-authored-by: Martin Balao <mbalao@redhat.com>
Co-authored-by: Francisco Ferrari Bihurriet <fferrari@redhat.com>